Skywire: Changed network mode to a property
I'm hoping doing:

    skywire = Skywire()
    skywire.networkMode = NetworkMode.On

is more intuitive than using setNetworkMode() and will make discovering all of
the functionality available in a Skywire easier. I.e. you can look up modules
within a Skywire and not have to keep track of the individual setX() and getX()
APIs.

Also changed the name of the AT interface to just 'at'. 'atInterface' is a
little heavy, and 'at' will fit a little better with incoming sub-module design
changes.