AT Interface: Added a helper for sending commands
This first phase will need to be accessible in a bit. Additionally, this new
helper uses a few tactics to (hopefully) make sending a little better,
particularly sending 'AT' to lock the AT interface and prevent URCs from being
sent just before entering the command (or just being buffered and never read
previously) and being read back after the command is sent and being interpreted
as if they were a part of the command's response.

This also removes the previous _clear() operation we were doing. With the line
reading, this shouldn't be necessary anymore, and it also would clobber any
futher improvements to URC handling.