AT Response: Added newline options for creation and parsing
We shuold offer a more live way of defining a modem's line ending style, since
not all modules are created equally. There's a circular import problem with the
master living in the AT interface, so the response will now define what it
considers the 'default', but will allow users to use something else.