Skywire: Organized files like the Nano
I'm still not a fan of __init__.py containing anything, since that makes it less
obvious how the module/packing is supposed to work, but I also don't have better
ideas for how to handle multiple files making up a package, where the files
would be organized under a directory with the same name as the base module (i.e.
skywire/ and skywire.py in the same directory). But this should still be pretty
Pythonic.